refactor(tests): eip-2537
