+++ tests/prague/eip2537_bls_12_381_precompiles/conftest.py
@@ -1,0 +1,148 @@
+"""
+Shared pytest definitions local to EIP-2537 tests.
+"""
+from typing import SupportsBytes
+
+import pytest
+from ethereum.crypto.hash import keccak256
+
+from ethereum_test_tools import Storage, TestAddress, Transaction
+from ethereum_test_tools.vm import Opcodes as Op
+
+from .spec import GAS_CALCULATION_FUNCTION_MAP
+
+###
+# Fixtures
+###
+
+
+@pytest.fixture
+def precompile_gas(precompile_address: int, input: bytes) -> int:
+    """Gas cost for the precompile."""
+    return GAS_CALCULATION_FUNCTION_MAP[precompile_address](len(input))
+
+
+@pytest.fixture
+def precompile_gas_modifier() -> int:
+    """
+    Used to modify the gas passed to the precompile, for testing purposes.
+
+    By default the call is made with the exact gas amount required for the given opcode,
+    but when this fixture is overridden, the gas amount can be modified to, e.g., test
+    a lower amount and test if the precompile call fails.
+    """
+    return 0
+
+
+@pytest.fixture
+def call_opcode() -> Op:
+    """
+    Type of call used to call the precompile.
+
+    By default it is Op.CALL, but it can be overridden in the test.
+    """
+    return Op.CALL
+
+
+@pytest.fixture
+def call_contract_post_storage() -> Storage:
+    """
+    Storage of the test contract after the transaction is executed.
+    Note: Fixture `call_contract_code` fills the actual expected storage values.
+    """
+    return Storage()
+
+
+@pytest.fixture
+def call_contract_code(
+    precompile_address: int,
+    precompile_gas: int,
+    precompile_gas_modifier: int,
+    expected_output: bytes | SupportsBytes,
+    call_opcode: Op,
+    call_contract_post_storage: Storage,
+) -> bytes:
+    """Code of the test contract."""
+    expected_output = bytes(expected_output)
+
+    # Depending on the expected output, we can deduce if the call is expected to succeed or fail.
+    call_succeeds = len(expected_output) > 0
+
+    assert call_opcode in [Op.CALL, Op.CALLCODE, Op.DELEGATECALL, Op.STATICCALL]
+    value = [0] if call_opcode in [Op.CALL, Op.CALLCODE] else []
+
+    code = (
+        Op.CALLDATACOPY(0, 0, Op.CALLDATASIZE())
+        + Op.SSTORE(
+            call_contract_post_storage.store_next(call_succeeds),
+            call_opcode(
+                precompile_gas + precompile_gas_modifier,
+                precompile_address,
+                *value,  # Optional, only used for CALL and CALLCODE.
+                0,
+                Op.CALLDATASIZE(),
+                0,
+                0,
+            ),
+        )
+        + Op.SSTORE(
+            call_contract_post_storage.store_next(len(expected_output)),
+            Op.RETURNDATASIZE(),
+        )
+        + Op.RETURNDATACOPY(0, 0, Op.RETURNDATASIZE())
+        + Op.SSTORE(
+            call_contract_post_storage.store_next(keccak256(expected_output)),
+            Op.SHA3(0, Op.RETURNDATASIZE()),
+        )
+    )
+    return code
+
+
+@pytest.fixture
+def call_contract_address() -> int:
+    """Code of the test contract."""
+    return 0xAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAA
+
+
+@pytest.fixture
+def pre(call_contract_address: int, call_contract_code: bytes):
+    """Code of the test contract."""
+    return {
+        call_contract_address: {
+            "balance": 0,
+            "nonce": 1,
+            "code": call_contract_code,
+        },
+        TestAddress: {
+            "balance": 1_000_000_000_000_000,
+            "nonce": 0,
+        },
+    }
+
+
+@pytest.fixture
+def post(call_contract_address: int, call_contract_post_storage: Storage):
+    """Test expected post outcome."""
+    return {
+        call_contract_address: {
+            "storage": call_contract_post_storage,
+        },
+    }
+
+
+@pytest.fixture
+def tx_gas_limit(precompile_gas: int) -> int:
+    """
+    Transaction gas limit used for the test (Can be overridden in the test).
+    """
+    return 10_000_000 + precompile_gas
+
+
+@pytest.fixture
+def tx(input: bytes, tx_gas_limit: int, call_contract_address: int) -> Transaction:
+    """Transaction for the test."""
+    return Transaction(
+        gas_limit=tx_gas_limit,
+        input=input,
+        to=call_contract_address,
+    )
