+++ tests/prague/eip2537_bls_12_381_precompiles/test_bls12_pairing.py
@@ -1,0 +1,19 @@
+"""
+abstract: Tests BLS12_PAIRING precompile of [EIP-2537: Precompile for BLS12-381 curve operations](https://eips.ethereum.org/EIPS/eip-2537)
+    Tests BLS12_PAIRING precompile of [EIP-2537: Precompile for BLS12-381 curve operations](https://eips.ethereum.org/EIPS/eip-2537).
+"""  # noqa: E501
+
+import pytest
+
+from ethereum_test_tools import Environment, StateTestFiller, Transaction
+
+from .helpers import vectors_from_file
+from .spec import FORK, Spec, ref_spec_2537
+
+REFERENCE_SPEC_GIT_PATH = ref_spec_2537.git_path
+REFERENCE_SPEC_VERSION = ref_spec_2537.version
+
+pytestmark = [
+    pytest.mark.valid_from(str(FORK)),
+    pytest.mark.parametrize("precompile_address", [Spec.PAIRING], ids=[""]),
+]
